Move product rating photos code to separate module
