+++ src/products/product-ratings/product-ratings.controller.spec.ts
@@ -11,7 +11,6 @@ import { ProductRatingDto } from '../dto/product-rating.dto';
 import { User } from '../../users/entities/user.entity';
 import { NotFoundError } from '../../errors/not-found.error';
 import { ProductRatingPhoto } from '../entities/product-rating-photo.entity';
-import { LocalFilesService } from '../../local-files/local-files.service';
 import { SettingsService } from '../../settings/settings.service';
 
 describe('ProductRatingsController', () => {
@@ -30,16 +29,6 @@ describe('ProductRatingsController', () => {
         RepositoryMockService.getProvider(Product),
         RepositoryMockService.getProvider(ProductRatingPhoto),
         DtoGeneratorService,
-        {
-          provide: LocalFilesService,
-          useValue: {
-            savePhoto: jest.fn((v) => ({
-              path: v.path,
-              mimeType: v.mimetype,
-            })),
-            createPhotoThumbnail: jest.fn((v: string) => v + '-thumbnail'),
-          },
-        },
         {
           provide: SettingsService,
           useValue: {
