+++ src/products/product-ratings/product-ratings.service.spec.ts
@@ -10,7 +10,6 @@ import { ProductRatingDto } from '../dto/product-rating.dto';
 import { User } from '../../users/entities/user.entity';
 import { NotFoundError } from '../../errors/not-found.error';
 import { ProductRatingPhoto } from '../entities/product-rating-photo.entity';
-import { LocalFilesService } from '../../local-files/local-files.service';
 import { SettingsService } from '../../settings/settings.service';
 
 describe('ProductRatingsService', () => {
@@ -28,16 +27,6 @@ describe('ProductRatingsService', () => {
         RepositoryMockService.getProvider(Product),
         RepositoryMockService.getProvider(ProductRatingPhoto),
         DtoGeneratorService,
-        {
-          provide: LocalFilesService,
-          useValue: {
-            savePhoto: jest.fn((v) => ({
-              path: v.path,
-              mimeType: v.mimetype,
-            })),
-            createPhotoThumbnail: jest.fn((v: string) => v + '-thumbnail'),
-          },
-        },
         {
           provide: SettingsService,
           useValue: {
