+++ src/products/product-ratings/product-ratings.service.ts
@@ -1,4 +1,4 @@
-import { Injectable, StreamableFile } from '@nestjs/common';
+import { Injectable } from '@nestjs/common';
 import { Repository } from 'typeorm';
 import { ProductRating } from '../entities/product-rating.entity';
 import { InjectRepository } from '@nestjs/typeorm';
@@ -6,8 +6,6 @@ import { ProductRatingDto } from '../dto/product-rating.dto';
 import { Product } from '../entities/product.entity';
 import { NotFoundError } from '../../errors/not-found.error';
 import { User } from '../../users/entities/user.entity';
-import { LocalFilesService } from '../../local-files/local-files.service';
-import { ProductRatingPhoto } from '../entities/product-rating-photo.entity';
 import { SettingsService } from '../../settings/settings.service';
 
 @Injectable()
@@ -17,9 +15,6 @@ export class ProductRatingsService {
     private readonly productRatingsRepository: Repository<ProductRating>,
     @InjectRepository(Product)
     private readonly productRepository: Repository<Product>,
-    @InjectRepository(ProductRatingPhoto)
-    private readonly productRatingPhotosRepository: Repository<ProductRatingPhoto>,
-    private readonly localFilesService: LocalFilesService,
     private settingsService: SettingsService,
   ) {}
 
@@ -91,95 +86,5 @@ export class ProductRatingsService {
       product: { id: productId },
     });
     return true;
-  }
-
-  async getProductRatingPhoto(
-    productId: number,
-    productRatingId: number,
-    photoId: number,
-    thumbnail: boolean,
-  ): Promise<StreamableFile> {
-    if (
-      (await this.settingsService.getSettingValueByName(
-        'Product rating photos',
-      )) !== 'true'
-    ) {
-      throw new NotFoundError('product rating photo');
-    }
-
-    const ratingPhoto = await this.productRatingPhotosRepository.findOne({
-      where: {
-        id: photoId,
-        productRating: { id: productRatingId, product: { id: productId } },
-      },
-    });
-    if (!ratingPhoto) {
-      throw new NotFoundError('product rating photo', 'id', photoId.toString());
-    }
-
-    const photoPath = thumbnail ? ratingPhoto.thumbnailPath : ratingPhoto.path;
-
-    const mimeType = thumbnail ? 'image/jpeg' : ratingPhoto.mimeType;
-
-    return await this.localFilesService.getPhoto(photoPath, mimeType);
-  }
-
-  async addProductRatingPhoto(
-    productId: number,
-    productRatingId: number,
-    file: Express.Multer.File,
-  ): Promise<ProductRating> {
-    if (
-      (await this.settingsService.getSettingValueByName(
-        'Product rating photos',
-      )) !== 'true'
-    ) {
-      throw new NotFoundError('product rating');
-    }
-    const productRating = await this.productRatingsRepository.findOne({
-      where: { id: productRatingId, product: { id: productId } },
-    });
-    if (!productRating) {
-      throw new NotFoundError(
-        'product rating',
-        'id',
-        productRatingId.toString(),
-      );
-    }
-    const photo = new ProductRatingPhoto();
-    const { path, mimeType } = await this.localFilesService.savePhoto(file);
-    photo.path = path;
-    photo.mimeType = mimeType;
-    photo.thumbnailPath = await this.localFilesService.createPhotoThumbnail(
-      file.path,
-    );
-    productRating.photos.push(photo);
-    return this.productRatingsRepository.save(productRating);
-  }
-
-  async deleteProductRatingPhoto(
-    productId: number,
-    productRatingId: number,
-    photoId: number,
-  ): Promise<ProductRating> {
-    if (
-      (await this.settingsService.getSettingValueByName(
-        'Product rating photos',
-      )) !== 'true'
-    ) {
-      throw new NotFoundError('product rating photo');
-    }
-    const productRating = await this.productRatingsRepository.findOne({
-      where: { id: productRatingId, product: { id: productId } },
-    });
-    if (!productRating) {
-      throw new NotFoundError(
-        'product rating',
-        'id',
-        productRatingId.toString(),
-      );
-    }
-    productRating.photos = productRating.photos.filter((p) => p.id !== photoId);
-    return await this.productRatingsRepository.save(productRating);
   }
 }
